+++ src/main/java/de/tum/in/www1/artemis/web/rest/SystemNotificationResource.java
@@ -16,8 +16,6 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import de.tum.in.www1.artemis.domain.notification.SystemNotification;
 import de.tum.in.www1.artemis.repository.SystemNotificationRepository;
-import de.tum.in.www1.artemis.security.annotations.EnforceNothing;
-import de.tum.in.www1.artemis.service.SystemNotificationService;
 import io.swagger.annotations.ApiParam;
 import tech.jhipster.web.util.PaginationUtil;
 import tech.jhipster.web.util.ResponseUtil;
@@ -31,11 +29,8 @@ public class SystemNotificationResource {
 
     private final SystemNotificationRepository systemNotificationRepository;
 
-    private final SystemNotificationService systemNotificationService;
-
-    public SystemNotificationResource(SystemNotificationRepository systemNotificationRepository, SystemNotificationService systemNotificationService) {
+    public SystemNotificationResource(SystemNotificationRepository systemNotificationRepository) {
         this.systemNotificationRepository = systemNotificationRepository;
-        this.systemNotificationService = systemNotificationService;
     }
 
     /**
@@ -65,19 +60,5 @@ public class SystemNotificationResource {
         log.debug("REST request to get SystemNotification : {}", id);
         Optional<SystemNotification> systemNotification = systemNotificationRepository.findById(id);
         return ResponseUtil.wrapOrNotFound(systemNotification);
-    }
-
-    /**
-     * Returns all system notifications with an expiry date in the future or no expiry date.
-     * This route is also accessible for unauthenticated users.
-     *
-     * @return the ResponseEntity with status 200 (OK) and with body the notification, or with status 404 (Not Found)
-     */
-    // TODO: /public
-    @GetMapping("system-notifications/active")
-    @EnforceNothing
-    public List<SystemNotification> getActiveAndFutureSystemNotifications() {
-        log.debug("REST request to get relevant system notifications");
-        return systemNotificationService.findAllActiveAndFutureSystemNotifications();
     }
 }
